+++ sale_usability/__init__.py
@@ -2,3 +2,4 @@
 
 from . import sale
 from . import product
+from . import partner
