+++ sale_usability/product.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# © 2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
+# © 2017 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from odoo import models, fields
@@ -11,3 +11,4 @@ class ProductTemplate(models.Model):
     track_service = fields.Selection(track_visibility='onchange')
     expense_policy = fields.Selection(track_visibility='onchange')
     invoice_policy = fields.Selection(track_visibility='onchange')
+    sale_line_warn = fields.Selection(track_visibility='onchange')
