+++ stock_usability/__init__.py
@@ -3,3 +3,4 @@
 from . import stock
 from . import procurement
 from . import product
+from . import partner
